+++ dfetch/vcs/git.py
@@ -233,6 +233,7 @@ class GitLocalRepo:
     """A git repository."""
 
     METADATA_DIR = ".git"
+    GIT_MODULES_FILE = ".gitmodules"
 
     def __init__(self, path: str | Path = ".") -> None:
         """Create a local git repo."""
@@ -258,7 +259,7 @@ class GitLocalRepo:
         src: str | None = None,
         must_keeps: list[str] | None = None,
         ignore: Sequence[str] | None = None,
-    ) -> str:
+    ) -> tuple[str, list[Submodule]]:
         """Checkout a specific version from a given remote.
 
         Args:
@@ -295,6 +296,14 @@ class GitLocalRepo:
             )
             run_on_cmdline(logger, ["git", "reset", "--hard", "FETCH_HEAD"])
 
+            run_on_cmdline(
+                logger,
+                ["git", "submodule", "update", "--init", "--recursive"],
+                env=_extend_env_for_non_interactive_mode(),
+            )
+
+            submodules = self.submodules()
+
             current_sha = (
                 run_on_cmdline(logger, ["git", "rev-parse", "HEAD"])
                 .stdout.decode()
@@ -304,7 +313,7 @@ class GitLocalRepo:
             if src:
                 self.move_src_folder_up(remote, src)
 
-            return str(current_sha)
+            return str(current_sha), submodules
 
     def move_src_folder_up(self, remote: str, src: str) -> None:
         """Move the files from the src folder into the root of the project.